Move request into a sub-module
